requirements.txt
Django==1.11.6
pytz==2017.2
Django==1.11.6
djangorestframework==3.7.1
pytz==2017.2
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
Django==1.11.6
djangorestframework==3.7.1
idna==2.6
pytz==2017.2
requests==2.18.4
urllib3==1.22
